requirements.txt
streamlit
python-dotenv
langchain
langchain-experimental
langchain-google-genai
google-generativeai
pandas
tabulate

